Implement remaning rbj filters
To simplify filter implementation, parameters used to update filter
coefficients were removed and are no longer checked if they are
within valid bounds. This means the user is responsible for
checking if parameters are within correct ranges to create a
working filter. Also, fixed the the module name from `rdj` to
`rbj`.

The new filters have not been tested for correctness, besides
lacking unit tests.